Auth for API works
This was a lot easier than I thought to use multiple
authentication mechanisms in the same app. Here we are
using both the HTTP basic for our API and a Local Strat
for web logins. They can both share the same method of
checking usernames and passwords, and they both set
req.user so that we can use that info in our routes.